pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wifi-watchdog"
version = "0.1.0"
authors = [ { name = "Your Name" } ]
license = { text = "MIT" }
description = "Raspberry Pi WiFi connectivity watchdog"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = ["pytest", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
pythonpath = ["src"]
